refactor: move transfer proxies into separate package
