+++ common/src/main/java/com/gregtechceu/gtceu/common/cover/ender_link/EnderFluidLinkCover.java
@@ -7,8 +7,8 @@ import com.gregtechceu.gtceu.api.cover.filter.FilterHandler;
 import com.gregtechceu.gtceu.api.cover.filter.FilterHandlers;
 import com.gregtechceu.gtceu.api.cover.filter.FluidFilter;
 import com.gregtechceu.gtceu.api.misc.FluidAmountHandler;
-import com.gregtechceu.gtceu.api.misc.LimitingFluidTransferProxy;
 import com.gregtechceu.gtceu.api.pipenet.enderlink.ITransferType;
+import com.gregtechceu.gtceu.api.transfer.proxies.LimitingFluidTransferProxy;
 import com.gregtechceu.gtceu.common.data.GTEnderLinkTransferTypes;
 import com.lowdragmc.lowdraglib.LDLib;
 import com.lowdragmc.lowdraglib.gui.widget.Widget;
